move kct package
