+++ core/src/test/java/com/klaytn/caver/feature/KIP17Test.java
@@ -3,19 +3,13 @@ package com.klaytn.caver.feature;
 import com.klaytn.caver.Caver;
 import com.klaytn.caver.crypto.KlayCredentials;
 import com.klaytn.caver.methods.response.KlayAccount;
-import com.klaytn.caver.methods.response.KlayAccountKey;
 import com.klaytn.caver.methods.response.KlayTransactionReceipt;
 import com.klaytn.caver.tx.gas.DefaultGasProvider;
-import com.klaytn.caver.tx.kct.KIP17;
-import com.klaytn.caver.tx.kct.KIP7;
+import com.klaytn.caver.kct.KIP17;
+import com.klaytn.caver.kct.KIP7;
 import com.klaytn.caver.tx.manager.TransactionManager;
-import com.klaytn.caver.utils.ChainId;
-import org.junit.Before;
 import org.junit.BeforeClass;
-import org.junit.Ignore;
 import org.junit.Test;
-import org.web3j.crypto.ECKeyPair;
-import org.web3j.crypto.Keys;
 import org.web3j.protocol.core.DefaultBlockParameterName;
 import org.web3j.tx.gas.StaticGasProvider;
 import org.web3j.utils.Numeric;
